Add PT2 Compile Events for triton and kernel compilation + load_by_key_path
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139402

Adds a few more dynamo_timed() to measure triton compilation and load_by_key_path times.

In the case of async compilation with multiple threads, we'll generate a single `kernel_compile` event that occurs when waiting on all the parallel compiles to finish.

In the case where async parallel compilation is disabled (or, compile threads are warming up), we'll generate a `triton_compile` event for each kernel.

The `triton_compile` events is a bit questionable: do we need a row for each triton compile event? It might eat up on our already low retention, so I might just remove that. Will discuss with @slarsen.
ghstack-source-id: 251405338
@exported-using-ghexport

Differential Revision: [D65215707](https://our.internmc.facebook.com/intern/diff/D65215707/)